Mise à jour des fichiers avant merge

+++ modules/Salariés/Chauffeur.cs
@@ -17,7 +17,7 @@ namespace Projet.Modules
         }
 
         [JsonConstructor]
-        public Chauffeur(string numeroSecuriteSociale, DateTime dateNaissance, DateTime dateEntreeSociete, string nom, string prenom, string adressePostale, string adresseMail, string telephone, string poste, decimal salaire, string managerNumeroSS)
+        public Chauffeur(string numeroSecuriteSociale, DateTime dateNaissance, DateTime dateEntreeSociete, string nom, string prenom, string adressePostale, string adresseMail, string telephone, string poste, double salaire, string managerNumeroSS)
          : base(numeroSecuriteSociale, dateNaissance, dateEntreeSociete, nom, prenom, adressePostale, adresseMail, telephone, poste, salaire, managerNumeroSS)
         {
             Poste = "Chauffeur";
