+++ modules/Salariés/Responsable.cs
@@ -14,7 +14,7 @@ namespace Projet.Modules
         }
 
         [JsonConstructor]
-        public Responsable(string numeroSecuriteSociale, DateTime dateNaissance, DateTime dateEntreeSociete, string nom, string prenom, string adressePostale, string adresseMail, string telephone, string poste, decimal salaire, string managerNumeroSS)
+        public Responsable(string numeroSecuriteSociale, DateTime dateNaissance, DateTime dateEntreeSociete, string nom, string prenom, string adressePostale, string adresseMail, string telephone, string poste, double salaire, string managerNumeroSS)
          : base(numeroSecuriteSociale, dateNaissance, dateEntreeSociete, nom, prenom, adressePostale, adresseMail, telephone, poste, salaire, managerNumeroSS)
         {
             Poste = "Responsable";
